+++ test/test_simulator.cpp
@@ -118,6 +118,18 @@ TEST_CASE("Set seed bit rate") {
   CHECK_NOTHROW(s.setSeedDeparture(seed));
 }
 
+TEST_CASE("Set seed src") {
+  Simulator s = Simulator();
+  unsigned int seed = 32;
+  CHECK_NOTHROW(s.setSeedSrc(seed));
+}
+
+TEST_CASE("Set seed dst") {
+  Simulator s = Simulator();
+  unsigned int seed = 32;
+  CHECK_NOTHROW(s.setSeedDst(seed));
+}
+
 TEST_CASE("Set goal connections") {
   Simulator s = Simulator();
   long long goalConnections = 1e7;
@@ -152,4 +164,112 @@ TEST_CASE("Uniform Variable Tests") {
     }
   }
   CHECK(flag);
+}
+
+TEST_CASE("Check decreasing Confidence Interval (Wald, 0.95)") {
+  float CI1, CI2, CI3;
+
+  Simulator s1 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s1);
+  s1.setGoalConnections(1e2);
+  s1.init();
+  s1.run();
+  CI1 = s1.waldCI();
+
+  Simulator s2 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s2);
+  s2.setGoalConnections(1e3);
+  s2.init();
+  s2.run();
+
+  CI2 = s2.waldCI();
+
+  Simulator s3 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s3);
+  s3.setGoalConnections(1e4);
+  s3.init();
+  s3.run();
+
+  CI3 = s3.waldCI();
+
+  CHECK(CI3 < CI2);
+  CHECK(CI2 < CI1);
+}
+
+TEST_CASE("Check decreasing Confidence Interval (Agresti-Coull, 0.95)") {
+  float CI1, CI2, CI3;
+
+  Simulator s1 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s1);
+  s1.setGoalConnections(1e2);
+  s1.init();
+  s1.run();
+  CI1 = s1.agrestiCI();
+
+  Simulator s2 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s2);
+  s2.setGoalConnections(1e3);
+  s2.init();
+  s2.run();
+
+  CI2 = s2.agrestiCI();
+
+  Simulator s3 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s3);
+  s3.setGoalConnections(1e4);
+  s3.init();
+  s3.run();
+
+  CI3 = s3.agrestiCI();
+
+  CHECK(CI3 < CI2);
+  CHECK(CI2 < CI1);
+}
+
+TEST_CASE("Check decreasing Confidence Interval (Wilson, 0.95)") {
+  float CI1, CI2, CI3;
+
+  Simulator s1 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s1);
+  s1.setGoalConnections(1e2);
+  s1.init();
+  s1.run();
+  CI1 = s1.wilsonCI();
+
+  Simulator s2 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s2);
+  s2.setGoalConnections(1e3);
+  s2.init();
+  s2.run();
+
+  CI2 = s2.wilsonCI();
+
+  Simulator s3 = Simulator(std::string("../networks/NSFNet.json"),
+                           std::string("../networks/NSFNet_routes.json"),
+                           std::string("../bitrate/bitrate.json"));
+  USE_ALLOC_FUNCTION(ExactFit, s3);
+  s3.setGoalConnections(1e4);
+  s3.init();
+  s3.run();
+
+  CI3 = s3.wilsonCI();
+
+  CHECK(CI3 < CI2);
+  CHECK(CI2 < CI1);
 }