+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2021-12-22
+
+### ADDED
+
+- Confidence intervals: Wilson, Wald, and Agresti-Coull
+- Added seed setters for the Src/Dst seeds' inside Simulator's class. 
+- Test cases inside test_simulator.cpp for the corresponding Src/Dst seed setters.
+
+### FIXED
+
+- There was a typo in JSON network file, and the related class: lenght -> length
+- Minor bugs
+
 ## [0.4.1] - 2021-07-28
 
 ### CHANGED
@@ -103,3 +116,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.0]: https://gitlab.com/DaniloBorquez/flex-net-sim/-/compare/v0.2.0...v0.3.0
 [0.4.0]: https://gitlab.com/DaniloBorquez/flex-net-sim/-/compare/v0.3.0...v0.4.0
 [0.4.1]: https://gitlab.com/DaniloBorquez/flex-net-sim/-/compare/v0.4.0...v0.4.1
+[0.5.0]: https://gitlab.com/DaniloBorquez/flex-net-sim/-/compare/v0.4.1...v0.5.0